Reorganize Store files
