+++ src/EventLogExpert.Store/EventLog/EventLogState.cs
@@ -4,7 +4,7 @@
 using EventLogExpert.Library.Models;
 using Fluxor;
 
-namespace EventLogExpert.Store.State;
+namespace EventLogExpert.Store.EventLog;
 
 /// <summary>
 ///     NOTE: Because Virtualize requires an ICollection<T>, we have to use
