+++ src/EventLogExpert.Store/StatusBar/StatusBarReducers.cs
@@ -1,11 +1,9 @@
 ﻿// // Copyright (c) Microsoft Corporation.
 // // Licensed under the MIT License.
 
-using EventLogExpert.Store.Actions;
-using EventLogExpert.Store.State;
 using Fluxor;
 
-namespace EventLogExpert.Store.Reducers;
+namespace EventLogExpert.Store.StatusBar;
 
 public class StatusBarReducers
 {
