+++ src/EventLogExpert.sln
@@ -13,7 +13,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventLogExpert.Test", "EventLogExpert.Test\EventLogExpert.Test.csproj", "{E5D3A2A2-6001-452F-8ECE-7263D72C9438}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EventLogExpert.Test", "EventLogExpert.Test\EventLogExpert.Test.csproj", "{E5D3A2A2-6001-452F-8ECE-7263D72C9438}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
