+++ src/EventLogExpert/Components/FilterPane.razor.cs
@@ -2,7 +2,8 @@
 // // Licensed under the MIT License.
 
 using EventLogExpert.Library.Models;
-using EventLogExpert.Store.Actions;
+using EventLogExpert.Store.EventLog;
+using EventLogExpert.Store.FilterPane;
 
 namespace EventLogExpert.Components;
