+++ src/EventLogExpert/_Imports.razor
@@ -8,7 +8,9 @@
 @using EventLogExpert.Library.Helpers
 @using EventLogExpert.Shared
 @using EventLogExpert.Store
-@using EventLogExpert.Store.State
+@using EventLogExpert.Store.EventLog
+@using EventLogExpert.Store.FilterPane
+@using EventLogExpert.Store.StatusBar
 
 @using Fluxor
 @using Fluxor.Blazor.Web.Components
